backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database & Auth
passlib>=1.7.4
bcrypt>=4.0.0
pyjwt>=2.8.0

# HTTP Client
httpx>=0.25.0
python-multipart>=0.0.9

# Excel Processing
openpyxl>=3.1.2
xlrd>=2.0.1

# ML & NLP (existing)
rank-bm25>=0.2.2
rapidfuzz>=3.5.0

# ML & NLP (new) - FIXED VERSION
sentence-transformers>=2.2.2
torch>=2.0.0
scikit-learn>=1.6.0
pandas>=2.1.0
numpy>=1.24.0
joblib>=1.3.0
iterative-stratification>=0.1.7

# ML Models
catboost>=1.2.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Data Processing & Encoding
category-encoders>=2.6.0

# AI/LLM
google-generativeai>=0.8.0